Migrate common component to spray
This is the first big step of migration to spray server. The common
bundle entirely relies on spray and not on tiscaf anymore.
The class hierarchy was pretty much simplified for the Rest API, there
is no split between synchronous/asynchronous or
standard/authenticated/role HLet.

The interface to CouchDB was also cleaned up and is available to any
person extending the Rest API.