Add publisher role and associated permissions
In a typical conference paper setting, the reviewers are anonymous and
only the review board knows who was assigned as reviewer to which paper.
However in a journal paper, the authors may know their reviewer but
cannot choose them.

This is why we introduce the `Publisher` role and two new permissions:
 - `manager-reviewers`,
 - `view-reviewers`

Depending on the paper kind, and the phase, these permissions may be
assigned to various roles.