Compute routes on demand
By using a lot of traits and mixin compositions, the order of
initializations is important and mixing `val`s, `def`s and `lazy val`s
is quite adventurous and leads to NPEs.
So now the routes of each module are computed once on demand, and
sub-routes are `def`s which avoids a lot of problems.