Try to fix error 405
Still unsuccessful, I have no idea whatsoever where it can come from,
looking at the code gives me no clue. The problem comes in the
`onSuccess` call when retrieving the current user, no error is thrown,
but it does not match the request either. There must be some good reason
for this, but I failed to find it until now.