Add default notification settings for new users
When the user registers with \BlueLaTeX, take over the default
notification settings from the configuration and create the appropriate
component.
This is not done using user hook because there is no need to have overly
modularized design where it is unnecessary. The core bundle can be seen
as a monolithic modules and need not using the module system.